Fix String (Reporter) InputBox errors on focusLost
