+++ netlogo-gui/src/main/window/InputBox.scala
@@ -5,6 +5,7 @@ package org.nlogo.window
 import org.nlogo.core.{ CompilerException, I18N }
 import org.nlogo.agent.InputBoxConstraint
 import org.nlogo.editor.AbstractEditorArea
+import org.nlogo.api.Exceptions
 import org.nlogo.api.Approximate.approximate
 import org.nlogo.api.Color.{getColor, getColorNameByIndex, modulateDouble}
 import org.nlogo.api.ModelReader.stripLines
@@ -31,7 +32,6 @@ abstract class InputBox(textArea:AbstractEditorArea, editDialogTextArea:Abstract
   protected val widgetLabel = new JLabel()
   protected var dialog: InputDialog = null
   private var _hasFocus = false
-  private var _returningFocus = false
   // grab the current editor kit from the editor area
   // everyone but string will use it but we need to
   // keep it around so we know what to set it to.
@@ -168,12 +168,8 @@ abstract class InputBox(textArea:AbstractEditorArea, editDialogTextArea:Abstract
             try inputText(inputType.readValue(InputBox.this.textArea.getText))
             catch {
               case ex@(_:LogoException|_:CompilerException|_:ValueConstraint.Violation) =>
-                println("erroring on focus lost")
-                if (! _returningFocus) showError(ex.asInstanceOf[Exception])
-                _returningFocus = true
                 org.nlogo.awt.EventQueue.invokeLater { () =>
                   InputBox.this.textArea.requestFocus()
-                  _returningFocus = false
                 }
             }
             editing = false
@@ -256,7 +252,6 @@ abstract class InputBox(textArea:AbstractEditorArea, editDialogTextArea:Abstract
     catch {
       case ex: LogoException => throw new IllegalStateException(ex)
       case ex@(_: CompilerException | _: ValueConstraint.Violation) =>
-        println("erroring editFinished")
         showError(ex.asInstanceOf[Exception])
     }
     true
@@ -390,7 +385,6 @@ abstract class InputBox(textArea:AbstractEditorArea, editDialogTextArea:Abstract
         }
         catch {
           case ex@(_:LogoException | _:CompilerException | _:ValueConstraint.Violation) =>
-            println("erroring on okAction")
             showError(ex.asInstanceOf[Exception])
         }
       }
@@ -410,7 +404,6 @@ abstract class InputBox(textArea:AbstractEditorArea, editDialogTextArea:Abstract
         try inputText(inputType.readValue(textArea1.getText))
         catch {
           case ex@(_:LogoException | _:CompilerException | _:ValueConstraint.Violation) =>
-            println("erroring on applyAction")
             showError(ex.asInstanceOf[Exception])
         }
       }
@@ -523,7 +516,9 @@ abstract class InputBox(textArea:AbstractEditorArea, editDialogTextArea:Abstract
     @throws(classOf[CompilerException])
     override def readValue(text: String) = {
       constraint.assertConstraint(text)
-      compiler.checkReporterSyntax(text)
+      Exceptions.ignoring(classOf[CompilerException]) {
+        compiler.checkReporterSyntax(text)
+      }
       text
     }
   }
