+++ netlogo-gui/src/test/headless/hubnet/TestClientInterface.scala
@@ -3,9 +3,13 @@
 package org.nlogo.headless.hubnet
 
 import org.nlogo.api.{ FileIO, LocalFile, ModelSection, ModelReader}
-import org.nlogo.util.TestUtils._
 import org.nlogo.headless.TestUsingWorkspace
 import org.nlogo.hubnet.protocol.ClientInterface
+import org.nlogo.util.ClassLoaderObjectInputStream
+
+import TestUtils._
+
+import org.scalatest.FunSuite
 
 class TestClientInterface extends TestUsingWorkspace {
 
@@ -33,5 +37,42 @@ class TestClientInterface extends TestUsingWorkspace {
   test("test roundTripSerialization method"){
     assert("s" == roundTripSerialization("s"))
     assert(7.asInstanceOf[AnyRef] == roundTripSerialization(7))
+  }
+}
+
+import java.io.{ByteArrayInputStream, ObjectOutputStream, ByteArrayOutputStream}
+
+// this object contains utility functions usable by other tests
+object TestUtils {
+
+  import org.scalatest.Assertions._
+
+  implicit class RoundTrip[T](t: T) {
+    def writeThenRead: T = roundTripSerialization(t)
+    def isSerializable = roundTripSerialization(t) === t
+  }
+
+  def roundTripSerialization[T](t: T) = {
+    val bytes = new ByteArrayOutputStream()
+    val out = new ObjectOutputStream(bytes)
+    out.writeObject(t)
+    out.flush()
+    val in = ClassLoaderObjectInputStream(
+      Thread.currentThread.getContextClassLoader,
+      new ByteArrayInputStream(bytes.toByteArray))
+    in.readObject().asInstanceOf[T]
+  }
+}
+
+// this class tests the utility functions in TestUtils.
+class TestUtilsTests extends FunSuite {
+  import TestUtils._
+
+  // fails in sbt unless we use ClassLoaderObjectInputStream instead of a regular ObjectInputStream
+  // in the roundTripSerialization.
+  // we think its because ObjectInputStream is using the bootstrap classloader. JC - 9/8/10
+  test("roundTripSerialization") {
+    assert("hello".isSerializable)
+    assert(7.isSerializable)
   }
 }
